Move the util method to a new util class
