+++ internal/venice-common/src/main/java/com/linkedin/venice/writer/VeniceWriter.java
@@ -2,6 +2,7 @@ package com.linkedin.venice.writer;
 
 import static com.linkedin.venice.ConfigKeys.INSTANCE_ID;
 import static com.linkedin.venice.ConfigKeys.LISTENER_PORT;
+import static com.linkedin.venice.writer.ChunkHelper.*;
 
 import com.linkedin.avroutil1.compatibility.AvroCompatibilityHelper;
 import com.linkedin.venice.annotation.Threadsafe;
@@ -35,10 +36,8 @@ import com.linkedin.venice.serialization.KeyWithChunkingSuffixSerializer;
 import com.linkedin.venice.serialization.VeniceKafkaSerializer;
 import com.linkedin.venice.serialization.avro.AvroProtocolDefinition;
 import com.linkedin.venice.serialization.avro.ChunkedValueManifestSerializer;
-import com.linkedin.venice.storage.protocol.ChunkId;
 import com.linkedin.venice.storage.protocol.ChunkedKeySuffix;
 import com.linkedin.venice.storage.protocol.ChunkedValueManifest;
-import com.linkedin.venice.utils.ByteUtils;
 import com.linkedin.venice.utils.ExceptionUtils;
 import com.linkedin.venice.utils.LatencyUtils;
 import com.linkedin.venice.utils.Time;
@@ -55,7 +54,6 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicLong;
-import java.util.function.BiFunction;
 import java.util.function.Supplier;
 import javax.annotation.Nonnull;
 import org.apache.avro.specific.FixedSize;
@@ -1084,7 +1082,7 @@ public class VeniceWriter<K, V, U> extends AbstractVeniceWriter<K, V, U> {
    * An interface which enables the key to contain parts of the {@param producerMetadata} within it, which is
    * useful for control messages and chunked values.
    */
-  private interface KeyProvider {
+  public interface KeyProvider {
     KafkaKey getKey(ProducerMetadata producerMetadata);
   }
 
@@ -1117,6 +1115,8 @@ public class VeniceWriter<K, V, U> extends AbstractVeniceWriter<K, V, U> {
         valueSchemaId,
         callback,
         reportSizeGenerator,
+        maxSizeForUserPayloadPerMessageInBytes,
+        keyWithChunkingSuffixSerializer,
         (keyProvider, putPayload) -> sendMessage(
             keyProvider,
             MessageType.PUT,
@@ -1133,6 +1133,8 @@ public class VeniceWriter<K, V, U> extends AbstractVeniceWriter<K, V, U> {
           valueSchemaId,
           callback,
           reportSizeGenerator,
+          maxSizeForUserPayloadPerMessageInBytes,
+          keyWithChunkingSuffixSerializer,
           (keyProvider, putPayload) -> sendMessage(
               keyProvider,
               MessageType.PUT,
@@ -1189,124 +1191,6 @@ public class VeniceWriter<K, V, U> extends AbstractVeniceWriter<K, V, U> {
         callback,
         leaderMetadataWrapper,
         Optional.of(logicalTs));
-  }
-
-  private void validateAvailableSizePerMessage(int sizeAvailablePerMessage, Supplier<String> sizeReport) {
-    if (sizeAvailablePerMessage < maxSizeForUserPayloadPerMessageInBytes / 2) {
-      /**
-       * If the key size is more than half the available payload per message, then we cannot even encode a
-       * {@link ChunkedValueManifest} with two keys in it, hence we would fail anyway. Might as well fail fast.
-       *
-       * N.B.: The above may not be strictly true, since Kafka-level compression may squeeze us under the limit,
-       * but this is not likely to work deterministically, so it would be iffy to attempt it on a best-effort
-       * basis. That being said, allowing keys to be as large as half the available payload size is probably
-       * overly permissive anyway. Ideally, the key size should be much smaller than this.
-       *
-       * TODO: Implement a proper key size (and value size) quota mechanism.
-       */
-      throw new VeniceException("Chunking cannot support this use case. The key is too large. " + sizeReport.get());
-    }
-  }
-
-  private ChunkedPayloadAndManifest chunksPayloadAndSend(
-      byte[] serializedKey,
-      byte[] payload,
-      int schemaId,
-      Callback callback,
-      Supplier<String> sizeReport,
-      BiFunction<KeyProvider, Put, Future<RecordMetadata>> sendMessageFunction) {
-    int sizeAvailablePerMessage = maxSizeForUserPayloadPerMessageInBytes - serializedKey.length;
-    validateAvailableSizePerMessage(sizeAvailablePerMessage, sizeReport);
-    int numberOfChunks = (int) Math.ceil((double) payload.length / (double) sizeAvailablePerMessage);
-    final ChunkedValueManifest chunkedValueManifest = new ChunkedValueManifest();
-    chunkedValueManifest.schemaId = schemaId;
-    chunkedValueManifest.keysWithChunkIdSuffix = new ArrayList<>(numberOfChunks);
-    chunkedValueManifest.size = payload.length;
-
-    KeyProvider keyProvider, firstKeyProvider, subsequentKeyProvider;
-    boolean chunkAwareCallback = callback instanceof ChunkAwareCallback;
-    ByteBuffer[] chunks = null;
-    if (chunkAwareCallback) {
-      // We only carry this state if it's going to be used, else we don't even instantiate this
-      chunks = new ByteBuffer[numberOfChunks];
-    }
-
-    /**
-     * This {@link ChunkedKeySuffix} instance gets mutated along the way, first in the loop, where its
-     * chunk index is incremented at each iteration, and then also on the first iteration of the loop,
-     * the {@link firstKeyProvider} will extract {@link ProducerMetadata} information out of the first
-     * message sent, to be re-used across all chunk keys belonging to this value.
-     */
-    final ChunkedKeySuffix chunkedKeySuffix = new ChunkedKeySuffix();
-    chunkedKeySuffix.isChunk = true;
-    chunkedKeySuffix.chunkId = new ChunkId();
-
-    subsequentKeyProvider = producerMetadata -> {
-      ByteBuffer keyWithSuffix =
-          ByteBuffer.wrap(keyWithChunkingSuffixSerializer.serializeChunkedKey(serializedKey, chunkedKeySuffix));
-      chunkedValueManifest.keysWithChunkIdSuffix.add(keyWithSuffix);
-      return new KafkaKey(MessageType.PUT, keyWithSuffix.array());
-    };
-    firstKeyProvider = producerMetadata -> {
-      chunkedKeySuffix.chunkId.producerGUID = producerMetadata.producerGUID;
-      chunkedKeySuffix.chunkId.segmentNumber = producerMetadata.segmentNumber;
-      chunkedKeySuffix.chunkId.messageSequenceNumber = producerMetadata.messageSequenceNumber;
-      return subsequentKeyProvider.getKey(producerMetadata);
-    };
-    for (int chunkIndex = 0; chunkIndex < numberOfChunks; chunkIndex++) {
-      int chunkStartByteIndex = chunkIndex * sizeAvailablePerMessage;
-      int chunkEndByteIndex = Math.min((chunkIndex + 1) * sizeAvailablePerMessage, payload.length);
-
-      /**
-       * We leave 4 bytes of headroom at the beginning of the ByteBuffer so that the Venice Storage Node
-       * can use this room to write the value header, without allocating a new byte array nor copying.
-       */
-      final int chunkLength = chunkEndByteIndex - chunkStartByteIndex;
-      byte[] chunkValue = new byte[chunkLength + ByteUtils.SIZE_OF_INT];
-      System.arraycopy(payload, chunkStartByteIndex, chunkValue, ByteUtils.SIZE_OF_INT, chunkLength);
-      ByteBuffer chunk = ByteBuffer.wrap(chunkValue);
-      chunk.position(ByteUtils.SIZE_OF_INT);
-
-      if (chunks != null) {
-        chunks[chunkIndex] = chunk;
-      }
-
-      Put putPayload = new Put();
-      putPayload.putValue = chunk;
-      putPayload.schemaId = AvroProtocolDefinition.CHUNK.getCurrentProtocolVersion();
-      putPayload.replicationMetadataVersionId = VENICE_DEFAULT_TIMESTAMP_METADATA_VERSION_ID;
-      putPayload.replicationMetadataPayload = EMPTY_BYTE_BUFFER;
-
-      chunkedKeySuffix.chunkId.chunkIndex = chunkIndex;
-      keyProvider = chunkIndex == 0 ? firstKeyProvider : subsequentKeyProvider;
-
-      try {
-        /**
-         * Here are the reasons to do a blocking call of 'sendMessage' with 'null' callback:
-         * 1. We don't want the upper layer know the chunking logic here. Right now, if we pass the callback parameter,
-         * it will cause the job failure because the message count of sent/completed in 'VeniceReducer' won't match;
-         * 2. Blocking call could guarantee correctness;
-         * 3. Infinite blocking means the following 'sendMessage' call will follow the config of the internal Kafka Producer,
-         * such as timeout, retries and so on;
-         */
-        sendMessageFunction.apply(keyProvider, putPayload).get();
-      } catch (Exception e) {
-        throw new VeniceException(
-            "Caught an exception while attempting to produce a chunk of a large value into Kafka... "
-                + getDetailedSizeReport(chunkIndex, numberOfChunks, sizeAvailablePerMessage, sizeReport),
-            e);
-      }
-    }
-    return new ChunkedPayloadAndManifest(chunkedValueManifest, chunks);
-  }
-
-  private String getDetailedSizeReport(
-      int chunkIndex,
-      int numberOfChunks,
-      int sizeAvailablePerMessage,
-      Supplier<String> sizeReport) {
-    return "Current chunk index: " + chunkIndex + ", " + "Number of chunks: " + numberOfChunks + ", "
-        + "Size available per message: " + sizeAvailablePerMessage + ", " + sizeReport.get();
   }
 
   private String getSizeReport(int serializedKeySize, int serializedValueSize, int replicationMetadataPayloadSize) {
@@ -1697,21 +1581,4 @@ public class VeniceWriter<K, V, U> extends AbstractVeniceWriter<K, V, U> {
     return maxSizeForUserPayloadPerMessageInBytes;
   }
 
-  class ChunkedPayloadAndManifest {
-    ByteBuffer[] payloadChunks;
-    ChunkedValueManifest chunkedValueManifest;
-
-    public ChunkedPayloadAndManifest(ChunkedValueManifest chunkedValueManifest, ByteBuffer[] payloadChunks) {
-      this.payloadChunks = payloadChunks;
-      this.chunkedValueManifest = chunkedValueManifest;
-    }
-
-    public ByteBuffer[] getPayloadChunks() {
-      return payloadChunks;
-    }
-
-    public ChunkedValueManifest getChunkedValueManifest() {
-      return chunkedValueManifest;
-    }
-  }
 }
